+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Updated translations ([#38](https://codeberg.org/silverpill/mitra-web/pulls/38)).
 - Ignore empty banner image.
+- Display `mention` notification with `reply` subtype as reply notification.
 
 ## [3.2.0] - 2024-08-18